requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
Pillow==10.4.0
whitenoise==6.7.0
python-dotenv==1.2.2
boto3==1.34.131
